% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Change_Password_Modal.R
\name{Change_Password_Modal}
\alias{Change_Password_Modal}
\title{Creates the Change Password Modal}
\usage{
Change_Password_Modal(status = NULL)
}
\value{
Modal Dialog in App
}
\description{
This function Creates the Change Password Modal.
}
\details{
Requires: shiny
}
\examples{

}
