% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copy_Item_Modal.R
\name{Copy_Item_Modal}
\alias{Copy_Item_Modal}
\title{Creates the Copy Inventory Item Modal}
\usage{
Copy_Item_Modal(copy_search_list)
}
\arguments{
\item{copy_search_list}{The copy search list for selection.}
}
\value{
Modal Dialog in App
}
\description{
This function Creates the Copy Inventory Item Modal.
}
\details{
Requires: shiny
}
\examples{

}
