% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item_Source_Modal_Edit.R
\name{Item_Source_Modal_Edit}
\alias{Item_Source_Modal_Edit}
\title{Creates Item Source Modal for Edit Inventory Item}
\usage{
Item_Source_Modal_Edit()
}
\value{
Modal Dialog in App
}
\description{
This function Creates the Item Source Modal.
}
\details{
Requires: shiny
}
\examples{

}
