% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item_Source_Modal_New.R
\name{Item_Source_Modal_New}
\alias{Item_Source_Modal_New}
\title{Creates Item Source Modal}
\usage{
Item_Source_Modal_New()
}
\value{
Modal Dialog in App
}
\description{
This function Creates the Item Source Modal.
}
\details{
Requires: shiny
}
\examples{

}
