% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Goal_Modal.R
\name{User_Goal_Modal}
\alias{User_Goal_Modal}
\title{Creates the User Goal Modal}
\usage{
User_Goal_Modal()
}
\value{
Modal Dialog in App
}
\description{
This function Creates the User Goal Modal.
}
\details{
Requires: shiny, shinyjs, shinyWidgets
}
\examples{

}
