% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Profile_Modal.R
\name{User_Profile_Modal}
\alias{User_Profile_Modal}
\title{Creates the User Profile Modal}
\usage{
User_Profile_Modal()
}
\value{
Modal Dialog in App
}
\description{
This function Creates the User Profile Modal.
}
\details{
Requires: shiny
}
